fix crash when cancelling a job
svn path=/trunk/kdesdk/cervisia/; revision=199362

+++ protocolview.cpp
@@ -128,7 +128,10 @@ bool ProtocolView::startJob()
 
 void ProtocolView::cancelJob()
 {
-    childproc->kill();
+    if( childproc )
+        childproc->kill();
+    else
+        job->cancel();
 }
